refactor: decompress
